requirements.txt
tensorflow-gpu==2.10
pandas==1.3.5
scikit-learn==1.0.2
matplotlib==3.5.3